requirements.txt
python-dotenv==1.0.0
pymysql==0.9.3
mysql-replication==0.23
csvkit==1.0.7
paramiko==3.3.1
cryptography==41.0.5
gzip-reader==0.1.3
pytz==2023.3